% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_tf.R
\name{agg_tf}
\alias{agg_tf}
\title{Calculate total fertility aggregates}
\usage{
agg_tf(
  dt,
  id_cols,
  age_mapping,
  missing_dt_severity = "stop",
  overlapping_dt_severity = "stop",
  present_agg_severity = "stop",
  na_value_severity = "stop",
  quiet = FALSE
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
ASFR input data. Must include all columns in \code{id_cols}, and a column for
'asfr'.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{age_mapping}{[\code{data.table()}]\cr
Specification of age interval to aggregate to. Required columns are
'age_start' and 'age_end'.}

\item{missing_dt_severity}{[\code{character(1)}]\cr
What should happen when \code{dt} is missing levels of \code{col_stem} that
prevent aggregation or scaling from occurring? Can be either 'skip',
'stop', 'warning', 'message', or 'none'. Default is 'stop'. See section on
'Severity Arguments' for more information.}

\item{overlapping_dt_severity}{[\code{character(1)}]\cr
When aggregating/scaling an interval variable or \code{collapse_interval_cols=TRUE}
what should happen when overlapping intervals are identified? Can be either
'skip', 'stop', 'warning', 'message', or 'none'. Default is 'stop'. See
section on 'Severity Arguments' for more information.}

\item{present_agg_severity}{[\code{logical(1)}]\cr
What should happen when \code{dt} already has requested aggregates (from
\code{mapping})? Can be either 'skip', 'stop', 'warning', 'message',
or 'none'. Default is 'stop'. See section on 'Severity Arguments' for more
information.}

\item{na_value_severity}{[\code{character(1)}]\cr
What should happen when 'NA' values are present in \code{value_cols}? Can be
either 'skip', 'stop', 'warning', 'message', or 'none'. Default is 'stop'.
See section on 'Severity Arguments' for more information.}

\item{quiet}{[\code{logical(1)}]\cr
Should progress messages be suppressed as the function is run? Default is
False.}
}
\value{
[\code{data.table()}]\cr Aggregated total fertility with columns for all
\code{id_cols} and a 'tf' column. Will only return the age groups specified in
\code{age_mapping}.
}
\description{
Calculate total fertility for a specified age range using
age-specific fertility rates.
}
\details{
Calculate total fertility aggregate for ages within a specific age range.
TFR (total fertility rate) is measured over the entire reproductive age span,
typically defined as between age 15 and 49 (or 10 and 54). \code{agg_tf} also
allows calculation of total fertility for other age spans like total
fertility under 25 and total fertility over 30.

Total fertility is calculated as the sum of ASFR multiplied by the number of
years in an age group. This number represents the average number of children
born to a woman if (1) she experiences a specific set of age specific
fertility rates and (2) she survives through the end of the age span.
Preston pg 95.

This is different from an age-specific fertility rate (ASFR) or a crude birth
rate (CBR), both of which are calculated as births/population for a
particular age group or all reproductive ages, respectively.
}
\section{Severity Arguments}{


\strong{\code{missing_dt_severity}}:

Check for missing levels of \code{col_stem}, the variable being aggregated or
scaled over.
\enumerate{
\item \code{stop}: throw error (this is the default).
\item \code{warning} or \code{message}: throw warning/message and continue with
aggregation/scaling for requested aggregations/scalings where expected input
data in \code{dt} is available.
\item \code{none}: don't throw error or warning, continue with aggregation/scaling
for requested aggregations/scalings where expected input data in \code{dt} is
available.
\item \code{skip}: skip this check and continue with aggregation/scaling.
}

\strong{\code{present_agg_severity}} (\code{agg} only):

Check for requested aggregates in \code{mapping} that are already present
\enumerate{
\item \code{stop}: throw error (this is the default).
\item \code{warning} or \code{message}: throw warning/message, drop aggregates and continue
with aggregation.
\item \code{none}: don't throw error or warning, drop aggregates and continue with
aggregation.
\item \code{skip}: skip this check and add to the values already present for the
aggregates.
}

\strong{\code{na_value_severity}}:

Check for 'NA' values in the \code{value_cols}.
\enumerate{
\item \code{stop}: throw error (this is the default).
\item \code{warning} or \code{message}: throw warning/message, drop missing values and
continue with aggregation/scaling where possible (this likely will cause
another error because of \code{missing_dt_severity}, consider setting
\code{missing_dt_severity = "skip"} for functionality similiar to \code{na.rm = TRUE}).
\item \code{none}: don't throw error or warning, drop missing values and continue
with aggregation/scaling where possible (this likely will cause another error
because of \code{missing_dt_severity}, consider setting
\code{missing_dt_severity = "skip"} for functionality similiar to \code{na.rm = TRUE}).
\item \code{skip}: skip this check and propagate \code{NA} values through
aggregation/scaling.
}

\strong{\code{overlapping_dt_severity}}:
Check for overlapping intervals that prevent collapsing to the most detailed
common set of intervals. Or check for overlapping intervals in \code{col_stem}
when aggregating/scaling.
\enumerate{
\item \code{stop}: throw error (this is the default).
\item \code{warning} or \code{message}: throw warning/message, drop overlapping intervals
and continue with aggregation/scaling where possible (this may cause another
error because of \code{missing_dt_severity}).
3 \code{none}: don't throw error or warning, drop overlapping intervals and
continue with aggregation/scaling where possible (this may cause another
error because of \code{missing_dt_severity}).
\item \code{skip}: skip this check and continue with aggregation/scaling.
}

}

\examples{
# calculate total fertility under 25 (ages 10 to 24)
dt <- data.table::data.table(
  age_start = c(10, 15, 20, 25, 30, 35, 40, 45),
  age_end = c(15, 20, 25, 30, 35, 40, 45, 50),
  asfr = c(0.00005, 0.02, 0.07, 0.08, 0.05, 0.02, 0.004, 0.0002)
)

dt <- agg_tf(
  dt = dt,
  id_cols = c("age_start", "age_end"),
  age_mapping = data.table::data.table(age_start = 10, age_end = 25)
)

}
\references{
Preston, Samuel, Patrick Heuveline, and Michel Guillot. 2001. Demography:
Measuring and Modeling Population. Wiley.
}
\seealso{
\code{\link[hierarchyUtils:agg_scale]{hierarchyUtils::agg()}}
}
