% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_popReconstruct.R
\docType{data}
\name{burkina_faso_data}
\alias{burkina_faso_data}
\alias{burkina_faso_initial_estimates}
\title{Burkina Faso example population data and ccmpp initial estimates}
\format{
\code{burkina_faso_data}: [\code{data.table()}] with 'population' data.
\itemize{
\item population: [\code{data.table()}] year-sex-age-specific census counts in
years after the baseline year (1975, 1985, 1995, 2005). Age groups are
five-year age groups from 0 to 80+.
}

\code{burkina_faso_initial_estimates}: list of [\code{data.table()}] of initial
estimates for each \code{\link[=ccmpp]{ccmpp()}} input. Calendar year intervals are for five-year
intervals between 1960 and 2005. Age groups are five-year age groups from 0
to 80+ (except 'survival' goes up to 85+).
\itemize{
\item srb: [\code{data.table()}] of year-specific sex ratio at birth estimates.
\item asfr: [\code{data.table()}] of year-age-specific average annual single-year
age-specific fertility rate estimates.
\item baseline: [\code{data.table()}] of year-sex-age specific baseline year
(1960) population counts.
\item survival: [\code{data.table()}] of year-sex-age-specific survivorship ratio
estimates.
\item net_migration: [\code{data.table()}] of year-sex-age-specific average annual
net migration proportion estimates
}

An object of class \code{list} of length 5.
}
\usage{
burkina_faso_data

burkina_faso_initial_estimates
}
\description{
Example data and initial estimates for reconstructing the female
population in Burkina Faso from 1960 to 2005.
}
\section{Possible [\code{data.table()}] columns for inputs}{


All contain:
\itemize{
\item \code{value_col}: [\code{numeric()}] contains value of interest for each input.
}

If year-specific:
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive).
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
}

If sex-specific:
\itemize{
\item sex: [\code{character()}] either 'female' or 'male'.
}

If age-specific:
\itemize{
\item age_start: [\code{integer()}] start of the age group (inclusive).
\item age_end: [\code{integer()}] end of the age group (exclusive).
}
}

\references{
Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2013. “Reconstructing Past Populations With Uncertainty From Fragmentary
Data.” Journal of the American Statistical Association 108 (501): 96–110.
\url{https://doi.org/10.1080/01621459.2012.737729}.

\href{https://cran.r-project.org/web/packages/popReconstruct/popReconstruct.pdf}{popReconstruct R Package}
}
\seealso{
\code{\link[=ccmpp]{ccmpp()}}, \code{\link[populationMethods:popReconstruct]{populationMethods::popReconstruct()}}
}
\keyword{datasets}
