% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbh.R
\name{calc_nqx_brass}
\alias{calc_nqx_brass}
\title{Estimate child mortality from sbh information}
\usage{
calc_nqx_brass(
  dt,
  id_cols = c("age_start", "age_end"),
  n_women_col = "n_women",
  ceb_col = "ceb",
  ced_col = "ced",
  model_schedule = "West"
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
sbh data with \code{id_cols}, \code{n_women_col}, \code{ceb_col}, \code{ced_col}}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}. This must include
'age_start' and 'age_end'.}

\item{n_women_col}{[\code{character(1)}]\cr
Name of column storing the number of women in each maternal age group.}

\item{ceb_col}{[\code{character(1)}]\cr
Name of column storing the number of children ever born alive in each
maternal age group.}

\item{ced_col}{[\code{character(1)}]\cr
Name of column storing the number of children ever born alive who have died
in each maternal age group.}

\item{model_schedule}{[\code{character(1)}]\cr
Name of model fertility schedule to use from \code{sbh_trussell_coeffs}. Default
is "West". Other implemented option is "North".}
}
\value{
[\code{data.table()}] with \code{id_cols}, the estimated probability of
dying ('nqx'), and the reference period ('tx').
}
\description{
Estimate child mortality from sbh information from one source
using the Brass method with Trussell coefficients.
}
\details{
The method reproduced in this function is described in Section B of Chapter
III in the UN Population Division Manual X. It estimates the probability of
child mortality with data classified by maternal age from one survey or
census.

Summary of steps:
\enumerate{
\item Calculate average parity for each maternal age group ('P(i)').
\item Calculate proportion of children who have died for each maternal age
group ('D(i)').
\item Use estimation equations to calculate probability of child mortality:
\deqn{k(i)=a(i) + b(i) (P(1)/P(2)) + c(i) (P(2)/P(3))}
\deqn{q(x) = k(i)D(i)}
\item Use estimation equation to calculate reference period:
\deqn{t(x)=a(i) + b(i) (P(1)/P(2)) + c(i) (P(2)/P(3))}
}
}
\examples{
calc_nqx_brass(
  dt = sbh_panama_1976,
  id_cols = c("sex", "age_start", "age_end"),
  model = "West"
)
calc_nqx_brass(
  dt = sbh_panama_1976,
  id_cols = c("sex", "age_start", "age_end"),
  model = "North"
)

}
\references{
UN Population Division. 1983. Manual X: Indirect Techniques for Demographic
Estimation. New York: United Nations, Department of Economic and Social
Affairs, ST/ESA/SER.A/81.
http://www.un.org/esa/population/techcoop/DemEst/manual10/manual10.html
}
