% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmpp.R
\name{ccmpp}
\alias{ccmpp}
\title{Cohort Component Method of Population Projection (CCMPP)}
\usage{
ccmpp(inputs, input_years, input_sexes, input_ages, value_col = "value")
}
\arguments{
\item{inputs}{list of [\code{data.table()}]s of initial estimates for each ccmpp
input. See \strong{Section: CCMPP inputs} and
\strong{Section: Possible [\code{data.table()}] columns for inputs} for more
information on each of the inputs.}

\item{input_years}{[\code{numeric()}]\cr
the unique 'year_start' column values in each of the \code{inputs}. The minimum
value must correspond to the only 'year_start' value in the baseline input.}

\item{input_sexes}{[\code{character()}]\cr
the unique 'sex' column values in each of the sex-specific \code{inputs}. Must
be either just 'female', or 'female' and 'male'.}

\item{input_ages}{[\code{numeric()}]\cr
the unique 'age_start' column values in each of the age-specific \code{inputs}.}

\item{value_col}{[\code{character(1)}]\cr
Name of the column containing the value of interest in each of the
\code{inputs}. Default is 'value'.}
}
\value{
[\code{data.table()}] of year-sex-age specific population counts.
}
\description{
Implements the deterministic female dominant cohort component
method of population projection.
}
\details{
Given information on the starting population size, the population at
later time points is a function of mortality, fertility, sex-ratio at birth,
and net-migration.

The cohort component method of population projection (CCMPP) projects an
sex-age specific population forward in time using leslie matrices
\code{\link[=leslie_matrix]{leslie_matrix()}} which encode information about fertility, sex-ratio at
birth, and mortality.

\deqn{Population(t + \delta) = Leslie[t, t + \delta] * Population(t)}

See \link{vignette("ccmpp")} or linked references for more details on this method.
}
\section{CCMPP inputs}{

\itemize{
\item srb: [\code{data.table()}] of year-specific sex ratio at birth estimates.
\item asfr: [\code{data.table()}] of year-age-specific average annual single-year
age-specific fertility rate estimates. Must include both maternal and
non-maternal age groups that are zero.
\item baseline: [\code{data.table()}] of year-sex-age specific baseline year
population counts.
\item survival: [\code{data.table()}] of year-sex-age-specific survivorship ratio
estimates.
\item net_migration: [\code{data.table()}] of year-sex-age-specific average annual
net migration proportion estimates
}
}

\section{Possible [\code{data.table()}] columns for inputs}{


All contain:
\itemize{
\item \code{value_col}: [\code{numeric()}] contains value of interest for each input.
}

If year-specific:
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive).
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
}

If sex-specific:
\itemize{
\item sex: [\code{character()}] either 'female' or 'male'.
}

If age-specific:
\itemize{
\item age_start: [\code{integer()}] start of the age group (inclusive).
\item age_end: [\code{integer()}] end of the age group (exclusive).
}
}

\examples{
population <- ccmpp(
  inputs = thailand_initial_estimates,
  input_years = seq(1960, 1995, 5),
  input_sexes = c("female", "male"),
  input_ages = seq(0, 80, 5)
)

}
\references{
Preston, Samuel, Patrick Heuveline, and Michel Guillot. 2001. Demography:
Measuring and Modeling Population. Wiley.
}
\seealso{
\code{\link[=leslie_matrix]{leslie_matrix()}} \link{vignette("ccmpp")}
}
