% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_lifetable.R
\name{check_mx_ax_qx}
\alias{check_mx_ax_qx}
\title{Check two of mx, ax, qx}
\usage{
check_mx_ax_qx(dt)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Data to check for mx, ax, and/or qx. Must
also have 'age_length' column.}
}
\value{
\code{dt} with input columns plus any of 'mx', 'ax', and 'qx' that is
missing in input. Or, returns error if input has fewer than two of these
parameters.
}
\description{
Helper function to check a data.table for two of mx, ax, and qx
and compute the missing parameter is one is missing.
}
\details{
Uses \code{\link[=mx_ax_to_qx]{mx_ax_to_qx()}}, \code{\link[=qx_ax_to_mx]{qx_ax_to_mx()}}, or \code{\link[=mx_qx_to_ax]{mx_qx_to_ax()}} function
to complete the set of three life table parameters.
}
