% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_dt_formatting.R
\name{matrix_to_dt}
\alias{matrix_to_dt}
\alias{dt_to_matrix}
\title{Convert between demographic data stored in data.table and matrix
format}
\usage{
matrix_to_dt(
  mdt,
  year_right_most_endpoint,
  age_right_most_endpoint = Inf,
  gen_end_interval_col = TRUE,
  value_col = "value",
  validate_arguments = TRUE
)

dt_to_matrix(
  dt,
  id_cols = c("year_start", "year_end", "sex", "age_start", "age_end"),
  value_col = "value",
  validate_arguments = TRUE
)
}
\arguments{
\item{mdt}{[\code{data.table()}]\cr
Input demographic data in the matrix format described in the details
section.}

\item{year_right_most_endpoint}{[\code{numeric(1)}]\cr
Assumed right most endpoint of the calendar year intervals. If \code{mdt} is
actually data collected at a specific point in time (like census data) and
not spanning intervals then assign 'NULL'.}

\item{age_right_most_endpoint}{[\code{numeric(1)}]\cr
Assumed right most endpoint of the age group intervals. Default is Inf.}

\item{gen_end_interval_col}{[\code{logical(1)}]\cr
Whether to automatically generate the end of the interval columns (like
'age'). Default is 'TRUE'.}

\item{value_col}{[\code{character(1)}]\cr
Name of the column containing the value of interest. Default is 'value'.}

\item{validate_arguments}{[\code{logical(1)}]\cr
Whether to validate that the input arguments are formatted correctly.
Default is 'TRUE'.}

\item{dt}{[\code{data.table()}]\cr
Input demographic data in the data.table format described in the details
section.}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}. This can only contain
year, sex, and age variables. 'year_start' and 'year_end' must be included,
and might contain 'sex' if the data is sex-specific, and 'age_start' and
'age_end' if the data is age-specific.}
}
\value{
\code{matrix_to_dt} returns a [\code{data.table()}] in data.table format as
described in the details section. \code{dt_to_matrix} returns a matrix of list of
matrices as described in the details section.
}
\description{
Convert demographic data (can be by age and/or sex) between two
data formats, [\code{data.table()}] and [\code{matrix()}]. When stored in matrix
form the data is required to be square (ie. same number of ages in every
year).
}
\details{
data.table format: When data is in data.table format then it must have a
column for 'year_start' and 'year_end', if the data is sex-specific it must
have a column for 'sex', and if the data is age-specific it must have
columns for 'age_start' and 'age_end.

matrix format: When data is in matrix format, columns represent the start of
each calendar year interval, rows represent the start of each age interval
and if the data is sex specific each sex will have a separate matrix stored
in a named list (names corresponding to each sex).
}
\examples{
output_matrix <- demCore:::dt_to_matrix(thailand_initial_estimates$survival)
output_dt <- demCore:::matrix_to_dt(output_matrix, year_right_most_endpoint = 2000)

}
