% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleLT}
\alias{exampleLT}
\title{Example life table}
\format{A data.table with columns:
\describe{
\item{age_start}{Integer, abridged ages 0, 1, 5, 10, ..., 95}
\item{age_end}{Integer, end of age group inclusive, 1, 5, 10, ... , 95, Inf}
\item{deaths}{Death counts}
\item{population}{Population counts}
\item{mx}{mortality rate (deaths / population)}
\item{ax}{mean person-years lived by those who die in interval. Obtained
using \code{mx_to_ax} function.}
\item{qx}{probability of death in interval conditional on survival to xth
birthday. Obtained using \code{mx_ax_to_qx} function.}
}}
\usage{
exampleLT
}
\description{
Example death count and population values. Also
includes mx (deaths / population), and ax and qx calculated using functions
from this package.
}
\examples{
exampleLT
}
\keyword{datasets}
