% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleLT2}
\alias{exampleLT2}
\title{Example life table 2}
\format{[\code{data.table()}] columns:
\describe{
\item{age_start}{[\code{integer()}]\cr Abridged ages 0, 1, 5, 10, ..., 85.
This is the 'x' variable in life table notation.}
\item{age_end}{[\code{integer()}]\cr End of age group inclusive, 1, 5, 10,
... , 85, Inf}
\item{age_length}{[\code{integer()}]\cr Width of the age interval in years.
This is the 'n' variable in life table notation.}
\item{deaths}{[\code{numeric()}]\cr Death counts in interval}
\item{population}{[\code{numeric()}]\cr Population counts in interval}
\item{mx}{[\code{numeric()}]\cr Mortality rate (deaths / population)}
\item{ax}{[\code{numeric()}]\cr Mean person-years lived by those who die in
interval.}
\item{qx}{[\code{numeric()}]\cr Probability of death in interval conditional
on survival to xth birthday.}
\item{px}{[\code{numeric()}]\cr Probability of survival in interval.}
\item{lx}{[\code{numeric()}]\cr Proportion surviving to age x.}
\item{dx}{[\code{numeric()}]\cr Proportion of cohort dying in age interval
x to x+n.}
\item{nLx}{[\code{numeric()}]\cr Person-years lived between x and x+n per
every one person in the cohort at birth.}
\item{Tx}{[\code{numeric()}]\cr Person-years lived beyond age x per every one
person in the cohort at birth.}
\item{ex}{[\code{numeric()}]\cr Life expectancy.}
}}
\usage{
exampleLT2
}
\description{
Example life table from Preston text book page 49. Post-hoc
adjustment has been made to use radix (l0) of 1 instead of 100,000, as
presented in the Preston book.

Original input data if from Austria, for males, from 1992, and comes
from the United Nations, 1994, as cited in Preston book.
}
\examples{
exampleLT2
}
\keyword{datasets}
