% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fNOR2010}
\alias{fNOR2010}
\title{Example life table (Norway 2010 females, from GBD 2019)}
\format{A data.table with columns:
\describe{
\item{age}{Integer, abridged ages 0, 1, 5, 10, ..., 95}
\item{location}{Character, "Norway"}
\item{deaths}{GBD 2019 death count estimate for Norway 2010 females}
\item{population}{GBD 2019 population estimate for Norway 2010 females}
\item{age_length}{Integer, length in years of an age group}
\item{mx}{mortality rate (deaths / population)}
\item{ax}{mean person-years lived by those who die in interval. Obtained
using \code{mx_to_ax} function.}
\item{qx}{probability of death in interval conditional on survival to xth
birthday. Obtained using \code{mx_ax_to_qx} function.}
}}
\source{
Vital registration data obtained from: ADD SOURCE HERE.
Census data obtained from: ADD SOURCE HERE.
See GBD methods as described in GBD 2019 Lancet paper and methods appendix
(cite GBD paper here).
}
\usage{
fNOR2010
}
\description{
GBD 2019 death count and population estimates, for females in
Norway in 2010, made using Norway vital registration and census data. Also
includes mx (deaths / population), and ax and qx calculated using functions
from this package.
}
\examples{
fNOR2010
}
\keyword{datasets}
