% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_lt_pars}
\alias{full_lt_pars}
\title{Regression parameters for abridged to full life table conversion}
\format{
[\code{data.table()}] columns:
\describe{
\item{sex}{[\code{character()}]\cr Sex the parameters correspond to: "male" or
"female"}
\item{age_start}{[\code{integer()}]\cr Single-year ages 1-109 representing the
age the parameters correspond to}
\item{age_end}{[\code{integer()}]\cr 1 + age_start}
\item{intercept}{[\code{numeric()}]\cr Intercept for the age- and sex-specific
regression model fit}
\item{slope}{[\code{numeric()}]\cr Slope for the age- and sex-specific
regression model fit}
}
}
\usage{
full_lt_pars
}
\description{
Parameters for age- and sex-specific regression model fit to
Human Mortality Database data, to predict single-year qx values from
abridged qx values:
\deqn{log(q_x) = intercept + slope*log(q_x,abridged)}
}
\examples{
full_lt_pars
}
\references{
Human Mortality Database. University of California, Berkeley (USA), and Max
Planck Institute for Demographic Research (Germany). Available at
www.mortality.org or www.humanmortality.de (data downloaded on Jan 2018).
}
\keyword{datasets}
