% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullLTpars}
\alias{fullLTpars}
\title{Regression parameters for abridged to full life table conversion}
\format{A data.table with columns:
\describe{
\item{sex}{Character, "male" or "female" for the sex the parameters
correspond to.}
\item{age}{Integer, single-year ages 1-109 representing the age the
parameters correspond to.}
\item{intercept}{Numeric, intercept for the age- and sex-specific regression
model fit.}
\item{slope}{Numeric, slope for the age- and sex-specific regression model
fit.}
}}
\source{
GBD 2019 methods (cite here).
Regression fit to Human Mortality Database life tables (cite here).
}
\usage{
fullLTpars
}
\description{
Parameters for age- and sex-specific regression model fit to
Human Mortality Database data, to predict single-year qx values from
abridged qx values:
\deqn{log(q_x) = intercept + slope*log(q_x,abridged)}
}
\examples{
fullLTpars
}
\keyword{datasets}
