% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ex.R
\name{gen_ex}
\alias{gen_ex}
\title{Calculate life expectancy (ex) from Tx and lx}
\usage{
gen_ex(dt, assert_na = T)
}
\arguments{
\item{dt}{data.table with \code{Tx} and \code{lx} variables}

\item{assert_na}{logical, default T, whether to check for NAs in generated ex}
}
\value{
None. Modifies the given data.table in-place.
}
\description{
Generate life expectancy (ex) at age x based on person-years lived above the
age group (Tx) and proportion of people alive at the start of age group (lx).
Given a data.table with \code{Tx} and \code{lx}, \eqn{e_x = T_x / l_x}.
}
\examples{
dt <- data.table::data.table(age = c(90),
                             Tx = c(3),
                             lx = c(0.1))
gen_ex(dt)

}
