% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lifetable_parameters.R
\name{gen_lifetable_parameters}
\alias{gen_lifetable_parameters}
\alias{gen_lx_from_qx}
\alias{gen_qx_from_lx}
\alias{gen_dx_from_lx}
\alias{gen_nLx}
\alias{gen_Tx}
\alias{gen_ex}
\title{Generate life table parameters}
\usage{
gen_lx_from_qx(dt, id_cols, assert_na = T)

gen_qx_from_lx(dt, id_cols, assert_na = T)

gen_dx_from_lx(dt, id_cols, assert_na = T)

gen_nLx(dt, id_cols, assert_na = T)

gen_Tx(dt, id_cols, assert_na = T)

gen_ex(dt, assert_na = T)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Life table(s). Includes columns 'age_start',
'age_end', and all \code{id_cols}. Additionally:
\itemize{
\item{\code{gen_lx_from_qx} requires 'qx'}
\item{\code{gen_qx_from_lx} requires 'lx'}
\item{\code{gen_dx_from_lx} requires 'lx'}
\item{\code{gen_nLx} requires 'lx', 'ax', 'dx'}
\item{\code{gen_Tx} requires 'nLx'}
\item{\code{gen_ex} requires 'lx', 'Tx'}
}}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify each row
of \code{dt}. Must include 'age_start' and 'age_end'.}

\item{assert_na}{[\code{logical()}]\cr Whether to check for NA values in the
generated variable.}
}
\value{
\code{dt} with column added for new life table parameter. Modifies
data.tables in place.
}
\description{
These functions perform life table calculations, to add selected life table
parameters from others, using established demographic relationships.
}
\details{
\strong{Parameter definitions:} \cr
For an age interval x to x+n:
\itemize{
\item mx = mortality rate (deaths / person-years)
\item qx = probability of death, conditional on survival to x-th birthday
\item ax = average years lived of those who died in the age interval
\item dx = number (or proportion of cohort) who died in the age interval
\item lx = number (or proportion of cohort) alive at the beginning of the age
interval
\item ex = life expectancy, expected value for years lived after age x
\item Tx =  total person-years lived beyond age x
\item nLx = total person-years lived in interval
}

\strong{gen_lx_from_qx:} Use probability of death (qx) to get the proportion of
survivors in the beginning of an age group (lx). l0 = 1; lx for ages > 0
is lx for previous age group times the probability of surviving the
previous age group.

\strong{gen_qx_from_lx:} Use survival proportion (lx) at age x and x+n to get
probability of death between x and x+n (qx). This relationship is an
algebraic equivalent to the \code{gen_lx_from_qx} relationship as described.
Terminal age qx is set to 1.

\strong{gen_dx_from_lx:} Calculate the proportion dying between age x and x+n (dx)
as the difference between the proportion surviving (lx) at age x and the
proportion surviving at age x+n. In the terminal age group all survivors
die in the age group, so dx = lx.

\strong{gen_nLx:} nLx is calculated as the sum of years lived by survivors and
years lived by those who die between ages x and x+n. The first component is
n times l(x+n). The second component is ax * dx. For terminal ages, assume
that average person years lived equals number at the start of the age group
divided by the mortality rate.

\strong{gen_Tx:} Person-years lived above age x (Tx) is the cumulative sum of
person-years lived by people in each age interval above age x (nLx).
For the terminal age group, Tx = nLx.

\strong{gen_ex:} Life expectancy (ex) at age x is based on person-years lived
above the age group (Tx) and proportion of people alive at the start of
age group (lx): ex = Tx / lx.
}
\examples{
dt <- data.table::data.table(
  sex = rep("both", 4),
  age_start = c(0, 5, 10, 15),
  age_end = c(5, 10, 15, Inf),
  age_length = c(5, 5, 5, Inf),
  mx = c(0.1, 0.2, 0.3, 0.4),
  ax = c(2.5, 2.5, 2.5, 2.5)
)
dt[, qx := mx_ax_to_qx(mx, ax, age_length)]
gen_lx_from_qx(dt, id_cols = c("sex", "age_start", "age_end"))
gen_dx_from_lx(dt, id_cols = c("sex", "age_start", "age_end"))
gen_nLx(dt, id_cols = c("sex", "age_start", "age_end"))
gen_Tx(dt, id_cols = c("sex", "age_start", "age_end"))
gen_ex(dt)
gen_qx_from_lx(dt, id_cols = c("sex", "age_start", "age_end"))

}
\seealso{
\itemize{
\item{\link[lifetableUtils:mx_qx_ax_conversions]{lifetableUtils::mx_qx_ax_conversions}}
\item{\link[lifetableUtils:lifetable]{lifetableUtils::lifetable}}
\item{\href{https://ihmeuw.github.io/lifetableUtils/articles/introduction_to_life_tables.html}{Introduction to life tables vignette}}
}
}
