% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship_ratio.R
\name{gen_lx_from_nLx_ax}
\alias{gen_lx_from_nLx_ax}
\title{Generate the lx life table parameter given nLx and ax.}
\usage{
gen_lx_from_nLx_ax(dt, id_cols)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data that includes columns for \code{id_cols}, 'nLx', and 'ax'.}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify each row
of \code{dt}. Must include 'age_start' and 'age_end'.}
}
\value{
\code{dt} with column added for new life table parameter. Modifies
data.tables in place.
}
\description{
Generate the lx life table parameter given nLx and ax.
}
\details{
Terminal age group:
\deqn{\begin{align} {}_{\infty}L_{x} &= n \cdot l_{x+n} + {}_{\infty}a_{x}
    \cdot {}_{\infty}d_{x} \\ &= {}_{\infty}a_{x} \cdot {}_{\infty}d_{x}
    \\ &= {}_{\infty}a_{x} \cdot l_{x} \end{align}}
\deqn{l_{x} = \frac{{}_{\infty}L_{x}}{{}_{\infty}a_{x}}}

Other age groups:
\deqn{\begin{align} {}_{n}L_{x} &= n \cdot l_{x+n} + {}_{n}a_{x} \cdot
    {}_{n}d_{x} \\ &= n \cdot l_{x+n} + {}_{n}a_{x} \cdot (l_x - l_{x+n})
    \end{align}}
\deqn{l_x = \frac{{}_{n}L_{x} - ((n - {}_{n}a_{x}) \cdot
    l_{x+n})}{{}_{n}a_{x}}}
}
\examples{
id_cols <- c("sex", "age_start", "age_end")
dt <- data.table::data.table(
  sex = rep("both", 4),
  age_start = c(0, 5, 10, 15),
  age_end = c(5, 10, 15, Inf),
  nLx = c(4, 2, 0.6, 0.1),
  ax = c(2.5, 2.5, 2.5, 2.5)
)
gen_lx_from_nLx_ax(dt, id_cols)

}
\seealso{
Other survivorship_ratio: 
\code{\link{gen_nLx_from_nSx}()},
\code{\link{nSx_from_lx_nLx_Tx}()}
}
\concept{survivorship_ratio}
