% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship_ratio.R
\name{gen_nLx_from_nSx}
\alias{gen_nLx_from_nSx}
\title{Generate nLx life table parameter from survivorship ratios}
\usage{
gen_nLx_from_nSx(dt, id_cols)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data that includes columns for \code{id_cols} and 'nSx'.}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify each row
of \code{dt}. Must include 'age_start' and 'age_end'.}
}
\value{
\code{dt} with column added for new life table parameter. Modifies
data.tables in place.
}
\description{
Generate the nLx life table parameter given survivorship ratios
(nSx).
}
\details{
See the \href{https://ihmeuw-demographics.github.io/demCore/index.html}{references page}
for the formatted equations below.

First age group:
\deqn{{}_{n}S_0 = \frac{{}_{n}L_{0}}{n \cdot l_0}}
\deqn{{}_{n}L_{0} = {}_{n}S_0 \cdot n \cdot l_0}

Other age groups:
\deqn{{}_{n}S_x = \frac{{}_{n}L_{x}}{{}_{n}L_{x-n}}}
\deqn{{}_{n}L_{x} = {}_{n}S_x \cdot {{}_{n}L_{x-n}}}

Terminal age group:
\deqn{{}_{n}S_x = \frac{T_{x}}{T_{x-n}} =
    \frac{{}_{\infty}L_{x}}{{}_{\infty}L_{x} + {}_{n}L_{x - n}}}
\deqn{{}_{\infty}L_{x} = \frac{{}_{n}S_x \cdot {}_{n}L_{x - n}}{1 - {}_{n}S_x}}
}
\examples{
id_cols <- c("sex", "age_start", "age_end")
dt <- data.table::data.table(
  sex = rep("both", 4),
  age_start = c(0, 5, 10, 15),
  age_end = c(5, 10, 15, Inf),
  nSx = c(0.95, 0.99, 0.99, 0.75)
)
gen_nLx_from_nSx(dt, id_cols)

}
\seealso{
Other survivorship_ratio: 
\code{\link{gen_lx_from_nLx_ax}()},
\code{\link{nSx_from_lx_nLx_Tx}()}
}
\concept{survivorship_ratio}
