% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_nrr.R
\name{gen_nrr}
\alias{gen_nrr}
\title{Generate Net Reproductive Rate (NRR)}
\usage{
gen_nrr(dt, id_cols, reproductive_age_start = 15, reproductive_age_end = 49)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data with asfr, sex-ratio at birth (\code{srb}), and nLx.}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify
observations in \code{dt}. Must include 'age_start', 'age_end', and 'sex'.}

\item{reproductive_age_start}{[\code{numeric(1)}]\cr
Age in years of start of reproductive age span (default 15 years).}

\item{reproductive_age_end}{[\code{numeric(1)}]\cr
Age in years of end of reproductive age span (default 49 years).}
}
\value{
dt [\code{data.table()}]\cr
Data table with \code{nrr} by \code{id_cols} (excluding 'age_start' and 'age_end').
}
\description{
Calculate Net Reproductive Rate (NRR) from X and X and
generate \code{nrr} variable in the data.table.
}
\examples{

}
