% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_summary_lt.R
\name{gen_summary_lt}
\alias{gen_summary_lt}
\title{Summarize life tables}
\usage{
gen_summary_lt(dt, id_cols, lt_params)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Draw-level life tables, columns include all
\code{id_cols}, all \code{lt_params}, and 'draw'.}

\item{id_cols}{[\code{character()}]\cr Column names in 'dt' that uniquely
identify all rows. Must include 'draw'.}

\item{lt_params}{[\code{character()}]\cr Life table parameters included in \code{dt}.
Example: c("mx", "ax", "dx"). Must include 'mx' and 'ax'.}
}
\value{
[\code{data.table()}]\cr Summarized life table with \code{id_cols},
'life_table_parameter', 'mean', 'lower', and 'upper'
}
\description{
Create mean and uncertainty intervals for summary-level life
tables based off of draws.
}
\details{
One draw is one independent simulation or calculation -- we
typically use 100 or 1000 to bootstrap uncertainty and propagate error
through multiple steps. First draws are collapsed to mean, lower
(2.5th percentile), and upper (97.5th percentile). Then, we make sure
mean-level life table parameters are consistent with one another by
re-calculating every requested parameter other than mx and ax, using
standard methods documented elsewhere.
}
\examples{
library(data.table)
data("austria_1992_lt")
dt <- data.table::data.table()
for(d in 1:100){
 dt_new <- copy(austria_1992_lt)
 dt_new[, draw := d]
 dt_new[, mx := mx * rnorm(1, mean = 1, sd = 0.05)]
 dt_new[, ax := ax * rnorm(1, mean = 1, sd = 0.05)]
 dt_new[, qx := NULL]
 dt <- rbind(dt, dt_new, fill = TRUE)
}
dt <- dt[!is.na(age_start)]
dt <- dt[, .(age_start, age_end, draw, mx, ax)]
dt <- gen_summary_lt(dt, id_cols = c("age_start", "age_end", "draw"),
  lt_params = c("mx", "ax"))

}
