% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_u5_ax.R
\name{gen_u5_ax}
\alias{gen_u5_ax}
\title{Generate ax for ages under 5 years}
\usage{
gen_u5_ax(dt, id_cols)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr columns:
\itemize{
\item 'age_start' with values 0 (for under-1) and 1 (for 1-4). Other ages
permitted but will result in NA output for ax or unchanged ax if ax in
input \code{dt}.
\item 'age_end' with values 1 (for under-1) and 5 (for 1-4).
\item 'sex' where values must be 'male', 'female', or 'both'
\item 'mx' for mortality rate
\item All additional columns from 'id_cols'
}}

\item{id_cols}{[\code{character()}]\cr columns that uniquely identify
observations in \code{dt}. Must include 'age_start', 'age_end', and 'sex'.}
}
\value{
modifies \code{dt} in place with 'ax' column added on.
}
\description{
Add ax onto a data.table for ages <1 and 1-4 years, with
algorithm based on under-1 mx.
}
\details{
Takes a data.table with age_start, age_end, sex, and infant
mortality rate and adds a column 'ax'. Values from:

Preston Samuel H, Patrick H, Michel G. Demography: measuring and modeling
population processes. MA: Blackwell Publishing. 2001.\tabular{lllll}{
    \tab  \tab Males \tab  \tab Females \cr
   1a0: \tab  \tab  \tab  \tab  \cr
   If 1m0 >= 0.107 \tab  \tab 0.330 \tab  \tab 0.350 \cr
   If 1m0 < 0.107 \tab  \tab 0.045 + 2.684 * 1m0 \tab  \tab 0.053 + 2.800 * 1m0 \cr
    \tab  \tab  \tab  \tab  \cr
   4a1: \tab  \tab  \tab  \tab  \cr
   If 1m0 >= 0.107 \tab  \tab 1.352 \tab  \tab 1.361 \cr
   If 1m0 < 0.107 \tab  \tab 1.651 - 2.816 * 1m0 \tab  \tab 1.522 - 1.518 * 1m0 \cr
}
}
\examples{
dt <- data.table::data.table(
  age_start = c(0, 1),
  age_end = c(1, 5),
  mx = c(0.09, 0.12),
  sex = c("male", "male")
)
dt <- gen_u5_ax(dt, id_cols = c("age_start", "age_end", "sex"))
}
