% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_u5_ax.R
\name{gen_u5_ax_from_qx}
\alias{gen_u5_ax_from_qx}
\title{Generate ax for ages under 5 years, from qx}
\usage{
gen_u5_ax_from_qx(dt, id_cols)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Columns:
\itemize{
\item 'age_start' with values 0 (for under-1) and 1 (for 1-4). Other ages
permitted but will result in NA output for ax or unchanged ax if ax in
input \code{dt}.
\item 'age_end' with values 1 (for under-1) and 5 (for 1-4).
\item 'sex' where values must be 'male', 'female', or 'both'
\item 'qx' for mortality rate
\item All additional columns from 'id_cols'
}}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify
observations in \code{dt}. Must include 'age_start', 'age_end', and 'sex'.}
}
\value{
Modifies \code{dt} in place with 'ax' column added on.
}
\description{
Add ax onto a data.table for ages <1 and 1-4 years, with
algorithm based on under-1 qx.
}
\details{
Takes a \code{data.table} with 'age_start', 'age_end', 'sex', and
infant probability of death ('qx') and adds a column 'ax'. There is no
absolute relationship between ax and qx without mx also known. Instead,
this function performs an inverse calculation of the calculation described
for \code{\link[demCore:gen_u5_ax]{demCore::gen_u5_ax()}} which estimates ax from 1m0. Since the inverse
involves a quadratic, we use \code{\link[stats:uniroot]{stats::uniroot()}} to find the solution.
}
\examples{
dt <- data.table::data.table(
  age_start = c(0, 1),
  age_end = c(1, 5),
  qx = c(0.0846, 0.3658),
  sex = c("male", "male")
)
gen_u5_ax_from_qx(dt, id_cols = c("age_start", "age_end", "sex"))
}
\references{
Preston Samuel H, Patrick H, Michel G. Demography: measuring and modeling
population processes. MA: Blackwell Publishing. 2001.
}
\seealso{
\code{\link[demCore:gen_u5_ax]{demCore::gen_u5_ax()}}
}
