% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_ax.R
\name{iterate_ax}
\alias{iterate_ax}
\title{Ax iteration method}
\usage{
iterate_ax(dt, id_cols, n_iterations = 30L, threshold = 0.01, quiet = F)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Life tables with columns: all \code{id_cols},
'age_start', 'age_end', 'dx', 'ax', 'qx', 'mx'.}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify each row
of \code{dt}. Must include 'age_start' and 'age_end'.}

\item{n_iterations}{[\code{integer(1)}]\cr Maximum number of iterations to run.}

\item{threshold}{[\code{numeric(1)}]\cr How far apart should ax and the
implied ax from dx be before removal from iteration. Default: 0.01.}

\item{quiet}{[\code{logical(1)}]\cr Whether to omit messages about iteration
progress. Default is F and progress messages are given.}
}
\value{
[\code{data.table()}]\cr Input life tables with ax, dx, qx modified.
Input mx is unchanged.
}
\description{
Implement iteration method to minimize the difference across
initialized ax and its expected relationship with dx. This method can be
used to start with naive values of ax (like age_length/2) and arrive
at more informed ax values based on an approximation that assumes a
relationship between ax and the age pattern of deaths (dx) in a cohort.
Since dx is often calculated using the mx --> ax --> qx method in the
first place, initialization and iteration is required.
}
\details{
Limitations: this method relies on initialized ax and dx values, and cannot
produce estimates for the first and last age groups. Preston pg 45.

Note: this is not the best method for ages under 5 years. See
\code{\link[=gen_u5_ax]{gen_u5_ax()}}.
}
\examples{
data("austria_1992_lt")
dt <- austria_1992_lt[,.(age_start, age_end, age_length, mx, ax, qx, dx, lx)]
new_dt <- iterate_ax(dt, id_cols = c("age_start", "age_end"))

\dontrun{
  # plot change in qx, ax, lx, dx, mx
  dt <- merge(dt, new_dt, by = c("age_start", "age_end"))
  plot(data = dt, qx.y/qx.x ~ age_start)
  plot(data = dt, ax.y/ax.x ~ age_start)
  plot(data = dt, lx.y/lx.x ~ age_start)
  plot(data = dt, dx.y/dx.x ~ age_start)
  plot(data = dt, mx.y/mx.x ~ age_start) # expect no change in mx
}

}
\references{
Preston Samuel H, Patrick H, Michel G. Demography: measuring and modeling
population processes. MA: Blackwell Publishing. 2001.
}
\seealso{
Preston Demography book pg 45. Also, see more details on this
method explained in the introduction to life tables vignette:
\code{vignette("introduction_to_life_tables", package = "demCore")}
}
