% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmpp.R
\name{leslie_matrix}
\alias{leslie_matrix}
\title{Make Leslie matrix}
\usage{
leslie_matrix(survival, asfr, srb, n_ages, int, female)
}
\arguments{
\item{survival}{[\code{numeric(n_ages + 1)} or \code{numeric(n_ages)}]\cr
Survivorship ratio, the proportion of people aged x - \code{int} that will be
alive \code{int} years later in a stationary population.}

\item{asfr}{[\code{numeric(n_ages)}]\cr
Annual age specific fertility rates NOT yet multiplied by \code{int}. Must
include both reproductive and non-reproductive age groups that are zero.}

\item{srb}{[\code{numeric(1)}]\cr
Sex ratio at birth.}

\item{n_ages}{[\code{integer(1)}]\cr
Number of age groups.}

\item{int}{[\code{integer(1)}]\cr
Width of the age groups and projection intervals. Usually 1 or 5.}

\item{female}{[\code{logical(1)}]\cr
Whether making Leslie matrix for projection of the female population
(versus male population).}
}
\value{
[\code{matrix(n_ages, n_ages)}]
}
\description{
Constructs the Leslie matrix needed for cohort component method
of population projection \code{\link[=ccmpp]{ccmpp()}}.
}
\details{
The Leslie matrix encodes two key pieces of information:
\enumerate{
\item The survivorship ratio is included in the off diagonal to age the
population in each age group \eqn{a} at time \eqn{t} to age group
\eqn{a + int} at time \eqn{t + int}.
\item When projecting the female population, the first row includes information
about age-specific fertility rates, maternal survivorship ratios, birth
survivorship ratios and the sex-ratio at birth to calculate the initial
population size of the youngest age group at time \eqn{t + int}.
}

See the references for more information.
}
\examples{
leslie <- leslie_matrix(
  survival = thailand_initial_estimates$survival[year_start == 1960 &
                                                  sex == "female", value],
  asfr = c(rep(0, 3),
           thailand_initial_estimates$asfr[year_start == 1960, value],
           rep(0, 7)),
  srb = thailand_initial_estimates$srb[year_start == 1960, value],
  n_ages = 17, int = 5, female = TRUE
)

}
\references{
Preston, Samuel, Patrick Heuveline, and Michel Guillot. 2001. Demography:
Measuring and Modeling Population. Wiley.
}
