% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship_ratio.R
\name{lx_from_lxpn_nLx_ax}
\alias{lx_from_lxpn_nLx_ax}
\title{Helper function to back calculate lx for one age group}
\usage{
lx_from_lxpn_nLx_ax(lxpn, nLx, ax, age_int)
}
\arguments{
\item{lxpn}{[\code{numeric(1)}]\cr}

\item{nLx}{[\code{numeric(1)}]\cr}

\item{ax}{[\code{numeric(1)}]\cr}

\item{age_int}{[\code{integer(1)}]\cr}
}
\value{
[\code{numeric(1)}]\cr
}
\description{
Helper function to back calculate lx from the lx value of the
next oldest age group (lxpn which is short for \eqn{l_{x+n}}), and the nLx
and ax values at age x.
}
