% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx_to_dx.R
\name{lx_to_dx}
\alias{lx_to_dx}
\title{lx to dx}
\usage{
lx_to_dx(dt, id_cols, terminal_age = 110, assert_na = T)
}
\arguments{
\item{dt}{data.table columns \code{age}, \code{lx}, and all variables in id_cols.}

\item{id_cols}{character vector of id columns that uniquely identify each row of \code{dt}.}

\item{terminal_age}{numeric representing the last age in the dataset. Default: 110.}

\item{assert_na}{logical for whether to check for NA values in the qx variables.}
}
\value{
\code{dt} with column "dx" added.
}
\description{
Given a data.table with the proportion surviving to each age group (\code{lx}),
calculate the proportion dying in the age interval (\code{dx}).

\eqn{d_x = l_x - l_{x+n}}
In terminal age group, \eqn{d_x = l_x}
}
\examples{
dt <- data.table::data.table(
  sex = rep("both", 4),
  age = c(0, 5, 10, 15),
  lx = c(1, 0.9, 0.7, 0.2)
)
dt <- lx_to_dx(dt, id_cols = c("sex"), terminal_age = 15)

}
