% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship_ratio.R
\name{nSx_from_lx_nLx_Tx}
\alias{nSx_from_lx_nLx_Tx}
\title{Calculate survivorship ratios}
\usage{
nSx_from_lx_nLx_Tx(dt, id_cols, terminal_age)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data that includes columns for \code{id_cols}, 'lx', 'nLx' and 'Tx'.}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify each row
of \code{dt}. Must include 'age_start' and 'age_end'.}

\item{terminal_age}{[\code{integer(1)}]\cr
The terminal age group in the population that the calculated survivorship
ratios will be used to project in. This must be less than or equal to the
maximum 'age_start' in \code{dt}.}
}
\value{
[\code{data.table()}] with the \code{id_cols} plus a column for 'nSx' with the
survivorship ratio value.
}
\description{
Calculate survivorship ratios to be used in projecting
populations forward in time with \code{\link[=ccmpp]{ccmpp()}}. The survivorship ratios
represent the proportion of people in one age group that will survive into
the next age group given the age-specific mortality rates in a lifetable.
}
\details{
First age group:
\deqn{{}_{n}S_0 = \frac{{}_{n}L_{0}}{n \cdot l_0}}

Other age groups:
\deqn{{}_{n}S_x = \frac{{}_{n}L_{x}}{{}_{n}L_{x-n}}}

Terminal age group:
\deqn{{}_{n}S_x = \frac{T_{x}}{T_{x-n}}}
}
\examples{
id_cols <- c("sex", "age_start", "age_end")
dt <- data.table::data.table(
  sex = rep("both", 4),
  age_start = c(0, 5, 10, 15),
  age_end = c(5, 10, 15, Inf),
  mx = c(0.1, 0.2, 0.3, 0.4),
  ax = c(2.5, 2.5, 2.5, 2.5)
)
lifetable(dt, id_cols)
nSx <- nSx_from_lx_nLx_Tx(dt, id_cols, terminal_age = 15)

}
\seealso{
\code{\link[=ccmpp]{ccmpp()}}

Other survivorship_ratio: 
\code{\link{gen_lx_from_nLx_ax}()},
\code{\link{gen_nLx_from_nSx}()}
}
\concept{survivorship_ratio}
