% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_to_lx.R
\name{qx_to_lx}
\alias{qx_to_lx}
\title{Use qx to calculate lx}
\usage{
qx_to_lx(dt, id_cols, assert_na = T)
}
\arguments{
\item{dt}{data.table with variables: qx, age, and all id columns}

\item{id_cols}{character vector of column names (excluding "age") that uniquely identify rows}

\item{assert_na}{logical for whether to check for NA values in the generated lx variables}
}
\value{
dt, with "lx" column added
}
\description{
Convert probability of dying (qx) to proportion of survivors in the beginning of an age group (lx).
Given a data.table with a qx variable, calculate and generate lx.
Starting with lx = 1 at age 0, lx = lx_previous * (1-qx_previous).
}
\examples{
dt <- data.table::data.table(
  sex = c(rep("male", 6), rep("female", 6)),
  qx = c(rep(.1, 6), rep(.2, 6)),
  age = rep(c(0, 1, seq(5, 20, 5)), 2)
)
dt <- qx_to_lx(dt, id_cols = c("sex"))

}
