% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sbh.R
\docType{data}
\name{sbh_panama_1976}
\alias{sbh_panama_1976}
\alias{sbh_trussell_coeffs}
\title{Example children ever born and surviving data from a Panama 1976
survey.}
\format{
21 rows and 6 columns  [\code{data.table()}] with counts of children ever born
alive (ceb) and children ever born alive who died (ced) tabulated by 5-year
maternal age  groups between age 15 and 49 and the child's sex. Also
includes a column for the number of women in each age group (n_women).

28 rows and 9 columns [\code{data.table()}]
\itemize{
\item \code{model}: the Coale-Demeny model fertility schedule used to produce the
coefficients.
\item \code{type}: whether the coefficients are to be used for estimating the child
mortality level ('q(x)/D(x)') or reference period ('t(x)').
\item \code{age_start}: start of maternal age group (inclusive).
\item \code{age_end}: end of maternal age group (exclusive).
\item \code{age_start_nqx}: start of child age group the probability of dying
applies to (inclusive).
\item \code{age_start_nqx}: end of child age group the probability of dying applies
to (exclusive).
\item \code{a_i}: \code{a(i)} coefficient
\item \code{b_i}: \code{b(i)} coefficient
\item \code{c_i}: \code{c(i)} coefficient
}
}
\usage{
sbh_panama_1976

sbh_trussell_coeffs
}
\description{
Example children ever born and surviving data from a Panama 1976
survey as described in the UN Population Division Manual X table 49.

Trussell multipliers to use in the Brass sbh method to convert
parity ratios and proportion of children ever born who died into estimates of
probability of dying before certain child ages. As described in the UN
Population Division Manual X tables 47 and 48.
}
\references{
UN Population Division. 1983. Manual X: Indirect Techniques for Demographic
Estimation. New York: United Nations, Department of Economic and Social
Affairs, ST/ESA/SER.A/81.
http://www.un.org/esa/population/techcoop/DemEst/manual10/manual10.html

UN Population Division. 1983. Manual X: Indirect Techniques for Demographic
Estimation. New York: United Nations, Department of Economic and Social
Affairs, ST/ESA/SER.A/81.
http://www.un.org/esa/population/techcoop/DemEst/manual10/manual10.html
}
\seealso{
\code{\link[=calc_nqx_brass]{calc_nqx_brass()}}

\code{\link[=sbh_trussell_coeffs]{sbh_trussell_coeffs()}}

\code{\link[=calc_nqx_brass]{calc_nqx_brass()}}

\code{\link[=sbh_panama_1976]{sbh_panama_1976()}}
}
\keyword{datasets}
