% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_qx.R
\name{scale_qx}
\alias{scale_qx}
\title{Scale qx values over age}
\usage{
scale_qx(dt, id_cols, missing_dt_severity = "stop")
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Data to be aggregated.
\itemize{
\item Must only include columns 'qx' and those specified in \code{id_cols}
\item Must include 'age_start' and 'age_end' columns
}}

\item{id_cols}{[\code{character()}]\cr
ID columns that uniquely identify each row of \code{dt}.}

\item{missing_dt_severity}{[\code{character(1)}]\cr
What should happen when \code{dt} is missing levels of \code{col_stem} that
prevent aggregation or scaling from occurring? Can be either 'skip',
'stop', 'warning', 'message', or 'none'. Default is 'stop'. See section on
'Severity Arguments' for more information.}
}
\value{
[\code{data.table()}]\cr Scaled qx values, has \code{id_cols} and 'qx'
columns for all age groups.
}
\description{
Scale granular qx values to be consistent with aggregate age
interval qx values already present in the data.
}
\details{
Convert to px-space, scale up age-hierarchy so that granular px
values multiply to aggregate px values, convert back to qx-space.
\code{scale_qx()} is a wrapper for \code{\link[hierarchyUtils:agg_scale]{hierarchyUtils::scale()}}.

This function is unique to aggregating and scaling qx over age because of
the dependence between age and the definition of qx. Aggregation of qx
across sex, location, or other variables should be done by first aggregating
ax and mx as population-weighted means (can use \code{\link[hierarchyUtils:agg_scale]{hierarchyUtils::agg()}} and
\code{\link[hierarchyUtils:agg_scale]{hierarchyUtils::scale()}} directly), then recalculating qx from mx and ax
(see \code{\link[=mx_ax_to_qx]{mx_ax_to_qx()}}).
}
\examples{
# scale qx up single-year to abridged u5 hierarchy
dt <- data.table::data.table(
  sex = "male",
  qx = c(0.03, 0.015, 0.005, 0.004, 0.001, 0.03, 0.05),
  age_start = c(0, 1, 2, 3, 4, 1, 0),
  age_end = c(1, 2, 3, 4, 5, 5, 5)
)
dt <- scale_qx(dt, id_cols = c("sex", "age_start", "age_end"))

}
\seealso{
Vignette on scaling multiplicative aggregates in \code{hierarchyUtils}.
}
