% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_u5_ax.R
\name{solve_u5_ax_from_qx}
\alias{solve_u5_ax_from_qx}
\title{Helper solve function for \code{gen_u5_ax_from_qx}}
\usage{
solve_u5_ax_from_qx(dt)
}
\arguments{
\item{dt}{[\code{data.table()}] A single life table.}
}
\description{
Function which solves for zero. Uses relationship between
under-5 mx and ax as described in \code{\link[=gen_u5_ax_from_mx]{gen_u5_ax_from_mx()}}.
}
\keyword{internal}
