% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr.R
\name{tfr}
\alias{tfr}
\title{Calculate total fertility}
\usage{
tfr(dt, id_cols, age_lower, age_upper)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Data
\itemize{
\item Must only include columns 'asfr' and those specified in \code{id_cols}
\item Must include 'age_start' and 'age_end' columns
}}

\item{age_lower}{[\code{integer(1)}]\cr
Start age for total fertility calculation}

\item{age_upper}{[\code{integer(1)}]\cr
End age for total fertility calculation}
}
\value{
[\code{data.table()}]\cr Total fertility values, has \code{id_cols} and 'tfr'
columns.
}
\description{
Calculate total fertility for a specified age range based on
age-specific fertility rates
}
\details{
Calculate total fertility aggregate for ages within \verb{age_lower`` and  }age_upper` such as total fertility under 25 and total fertility over 30.
}
\examples{
# calculate total fertility under 25 (ages 10 to 24)
dt <- data.table('asfr' = c(0.00005, 0.02, 0.07, 0.08, 0.05, 0.02, 0.004, 0.0002),
                 'age_start' = c(10, 15, 20, 25, 30, 35, 40, 45),
                 'age_end' = c(15, 20, 25, 30, 35, 40, 45, 50))
dt <- tfr(test, c(), 10, 25)
}
