% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_popReconstruct.R
\docType{data}
\name{thailand_data}
\alias{thailand_data}
\alias{thailand_initial_estimates}
\title{Thailand example population data and ccmpp initial estimates}
\format{
\code{thailand_data}: [\code{data.table()}] with 'population' data.
\itemize{
\item population: [\code{data.table()}] year-sex-age-specific census counts in
years after the baseline year (1970, 1980, 1990, 2000). Age groups are
five-year age groups from 0 to 80+.
}

\code{thailand_initial_estimates}: list of [\code{data.table()}] of initial
estimates for each \code{\link[=ccmpp]{ccmpp()}} input. Calendar year intervals are for five-year
intervals between 1960 and 2000. Age groups are five-year age groups from 0
to 80+ (except 'survival' goes up to 85+). See \strong{Section: Inputs} for more
information on each of the inputs.

An object of class \code{list} of length 5.
}
\usage{
thailand_data

thailand_initial_estimates
}
\description{
Example data and initial estimates for reconstructing the female
and male population in Thailand from 1960 to 2000.
}
\section{Inputs}{

srb: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item \code{value_col}: [\code{numeric()}] sex-ratio at birth estimates.
}

asfr: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages_asfr' setting.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual age-specific fertility rate estimates.
}

baseline: [\code{data.table()}]\cr
\itemize{
\item year: [\code{integer()}] mid-year for population estimate.
Corresponds to 'years' \code{setting}.
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] baseline year population count estimates.
}

survival: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] survivorship ratio estimates.
}

net_migration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual net-migration proportion estimates.
}

immigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual immigration proportion estimates.
}

emigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual emigration proportion estimates.
}
}

\references{
Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2015. “Bayesian Reconstruction of Two-Sex Populations by Age: Estimating Sex
Ratios at Birth and Sex Ratios of Mortality.” Journal of the Royal
Statistical Society. Series A: Statistics in Society 178 (4): 977–1007.
\url{https://doi.org/10.1111/rssa.12104}.

\href{https://github.com/markalava/Bayesian-Reconstruction}{markalava/Bayesian-Reconstruction github repo}
}
\seealso{
\code{\link[=ccmpp]{ccmpp()}}
}
\keyword{datasets}
