% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_u5_ax.R
\name{validate_gen_u5_ax}
\alias{validate_gen_u5_ax}
\title{Helper for \code{gen_u5_ax} validations}
\usage{
validate_gen_u5_ax(dt, id_cols, param_cols)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr Columns:
\itemize{
\item 'age_start' with values 0 (for under-1) and 1 (for 1-4). Other ages
permitted but will result in NA output for ax or unchanged ax if ax in
input \code{dt}.
\item 'age_end' with values 1 (for under-1) and 5 (for 1-4).
\item 'sex' where values must be 'male', 'female', or 'both'
\item 'mx' for mortality rate
\item All additional columns from 'id_cols'
}}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify
observations in \code{dt}. Must include 'age_start', 'age_end', and 'sex'.}

\item{param_cols}{[\code{character()}] "mx" for \code{gen_u5_ax} and "qx" for
\code{gen_u5_ax_from_qx}}
}
\description{
Helper for \code{gen_u5_ax} validations
}
