% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_u5_ax.R
\name{zero_u5_ax_from_qx}
\alias{zero_u5_ax_from_qx}
\title{Helper objective function for \code{gen_u5_ax_from_qx}}
\usage{
zero_u5_ax_from_qx(ax, qx, A, B)
}
\arguments{
\item{ax}{[\code{numeric(1)}] average person-years lived by those who die (1a0)}

\item{qx}{[\code{numeric(1)}] probability of death (1q0)}

\item{A}{[\code{numeric(1)}] Intercept for 1m0 --> 1a0}

\item{B}{[\code{numeric(1)}] Slope for 1m0 --> 1a0}
}
\description{
Function we want to zero. Uses relationship between under-5
mx and ax as described in \code{\link[=gen_u5_ax_from_mx]{gen_u5_ax_from_mx()}}. \strong{Use only for 1a0.}
}
\keyword{internal}
