% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abridged_to_full.R
\name{abridged_to_full}
\alias{abridged_to_full}
\title{Generate full life table from abridged life table}
\usage{
abridged_to_full(
  dt,
  id_cols,
  regression_fits,
  regression_id_cols,
  terminal_age = 110,
  lx_spline_start_age = 15,
  lx_spline_end_age = 100,
  preserve_input_ax_ages = c(0, terminal_age)
)
}
\arguments{
\item{dt}{[\code{data.table()}] with variables age, all \code{id_cols}, 'qx', 'ax'}

\item{id_cols}{[\code{character()}] variables that uniquely identify rows,
must include 'age'}

\item{regression_fits}{[\code{data.table()}] with variables from \code{id_cols}, plus
'intercept', slope'}

\item{regression_id_cols}{[\code{character()}] variables that uniquely identify
regression parameters. Must include 'age' and be contained by \code{id_cols}.}

\item{terminal_age}{[\code{integer(1)}] max age that is being computed
(default: 110)}

\item{lx_spline_start_age}{[\code{integer(1)}] age (inclusive) to start using lx
spline rather than regression fits. Use 0 to use lx spline for all ages, or
integer > 110 or Inf to use regression results for all.}

\item{lx_spline_end_age}{[\code{integer(1)}] age (non-inclusive) to end spline
and begin using regression fits.}

\item{preserve_input_ax_ages}{[\code{integer()}] ages to preserve the input ax
values for. This is typically the first age group 0-1 and the terminal age
group 110+.}
}
\value{
data.table with columns id_cols, age, qx, ax
}
\description{
Convert abridged (5-year age group) life tables to full
(single-year-age) life tables using specified regression parameters or
lx spline
}
\details{
This function includes two different methods for expanding
abridged life tables to full life tables. The first method is a
\strong{monotonic cubic spline} over lx. Because lx is always decreasing, the
monotonic spline fits the general curve and pattern well, basically
splining between the knots at each abridged-year increment. However, the
spline can become wild and unreasonable in age groups with dramatic changes
in mortality rate over age.

The second method is a \strong{regression method} to predict log full qx from
log abridged qx:
\deqn{log(qx) = B0 + B1 log(qx,abridged)}
Where the parameters \eqn{B0} and \eqn{B1} are fit using high
quality single-year life tables (like those from Human Mortality Database),
and passed into this function. The regression-based method has its own
challenges, where scalloping qx patterns may arise because we do not take
consecutive abridged qx values into account simultaneously.

The function also allows for a combination of the methods to be used, with
separation on age. The default age cutoffs reflect the values we found to
work best: regression method for ages <15 and >100 and spline method for
ages between 15 and 100.

Values of ax are assumed to be 0.5 for all single-year ages, except those
passed into \code{preserve_input_ax_ages} argument. HMD also assumes 0.5 for
single-years, but we may be introducing an inconsistency between abridged
and full life table mx values that are calculated from these differing ax
values. Therefore, this is an area for future methods development.
}
\examples{
\dontrun{
data("fNOR2010")
data("fullLTpars")
id_cols <- c("location", "age")
dt <- abridged_to_full(dt = fNOR2010, regression_fits = fullLTpars,
  id_cols = id_cols, regression_id_cols = c("age"), terminal_age = 95)
}
}
