% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_qx.R
\name{agg_qx}
\alias{agg_qx}
\title{Aggregate qx values over age}
\usage{
agg_qx(dt, age_start, age_end, id_cols = c())
}
\arguments{
\item{dt}{data.table with variables: qx, age_start, age_end, all id_cols}

\item{age_start}{numeric, year of age to start at (e.g. 0 for 5q0)}

\item{age_end}{numeric, year of age to end at (e.g. 5 for 5q0)}

\item{id_cols}{character vector of column names in dt that, along with age_start
and age_end, uniquely identify all rows in dt}
}
\value{
data.table with id_cols and a variable called qx_#q#, where the first is
age_end - age_start, and the second is age_start
}
\description{
Given a data.table with a qx variable and ID variables that uniquely
identify the data, compile granular ages and aggregate into combined qx values.
}
\examples{
dt <- data.table::data.table(
  id = c(rep(1, 5), rep(2, 5)), qx = c(rep(.1, 5), rep(.2, 5)),
  age_start = rep(seq(15, 35, 5), 2),
  age_end = rep(seq(20, 40, 5), 2)
)
agg_qx(dt, age_start = 15, age_end = 40, id_cols = "id")

}
