% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_popReconstruct.R
\docType{data}
\name{burkina_faso_data}
\alias{burkina_faso_data}
\alias{burkina_faso_initial_estimates}
\title{Burkina Faso example population data and ccmpp initial estimates}
\format{\code{burkina_faso_data}: list containing sex-specific matrices of population
count data.
\itemize{
\item population: named list with female matrix of age-specific census
counts in years after the baseline year (1975, 1985, 1995, 2005). Rows
correspond to age groups while columns correspond to years.
}

\code{burkina_faso_initial_estimates}: list containing matrices of initial
estimates for each \code{\link[=ccmpp]{ccmpp()}} input.
\itemize{
\item srb: single row matrix of initial estimates of sex ratio at birth.
\item asfr: matrix of initial estimates of average annual single-year
age-specific fertility rates.
\item baseline: List with female single column matrix of age-specific
population counts in the baseline year (1960).
\item survival: List with female matrix of initial estimates of age-specific
survivorship ratios
\item net_migration: List with female matrix of initial estimates of
age-specific average annual net migration proportions.
}}
\usage{
burkina_faso_data

burkina_faso_initial_estimates
}
\description{
Example data and initial estimates for reconstructing the female
population in Burkina Faso from 1960 to 2005.
}
\references{
Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2013. “Reconstructing Past Populations With Uncertainty From Fragmentary
Data.” Journal of the American Statistical Association 108 (501): 96–110.
\url{https://doi.org/10.1080/01621459.2012.737729}.

\href{https://cran.r-project.org/web/packages/popReconstruct/popReconstruct.pdf}{popReconstruct R Package}
}
\seealso{
\code{\link[=ccmpp]{ccmpp()}}, \code{\link[populationMethods:popReconstruct]{populationMethods::popReconstruct()}}
}
\keyword{datasets}
