% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_popReconstruct.R
\docType{data}
\name{burkina_faso_data}
\alias{burkina_faso_data}
\alias{burkina_faso_initial_estimates}
\title{Burkina Faso example population data and ccmpp initial estimates}
\format{
\code{burkina_faso_data}: [\code{data.table()}] with 'population' data.
\itemize{
\item population: [\code{data.table()}] year-sex-age-specific census counts in
years after the baseline year (1975, 1985, 1995, 2005). Age groups are
five-year age groups from 0 to 80+.
}

\code{burkina_faso_initial_estimates}: list of [\code{data.table()}] of initial
estimates for each \code{\link[=ccmpp]{ccmpp()}} input. Calendar year intervals are for five-year
intervals between 1960 and 2005. Age groups are five-year age groups from 0
to 80+ (except 'survival' goes up to 85+). See \strong{Section: Inputs} for more
information on each of the inputs.

An object of class \code{list} of length 5.
}
\usage{
burkina_faso_data

burkina_faso_initial_estimates
}
\description{
Example data and initial estimates for reconstructing the female
population in Burkina Faso from 1960 to 2005.
}
\section{Inputs}{

srb: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item \code{value_col}: [\code{numeric()}] sex-ratio at birth estimates.
}

asfr: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages_asfr' setting.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual age-specific fertility rate estimates.
}

baseline: [\code{data.table()}]\cr
\itemize{
\item year: [\code{integer()}] mid-year for population estimate.
Corresponds to 'years' \code{setting}.
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] baseline year population count estimates.
}

survival: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] survivorship ratio estimates.
}

net_migration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual net-migration proportion estimates.
}

immigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual immigration proportion estimates.
}

emigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual emigration proportion estimates.
}
}

\references{
Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2013. “Reconstructing Past Populations With Uncertainty From Fragmentary
Data.” Journal of the American Statistical Association 108 (501): 96–110.
\url{https://doi.org/10.1080/01621459.2012.737729}.

\href{https://cran.r-project.org/web/packages/popReconstruct/popReconstruct.pdf}{popReconstruct R Package}
}
\seealso{
\code{\link[=ccmpp]{ccmpp()}}
}
\keyword{datasets}
