% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmpp.R
\name{ccmpp}
\alias{ccmpp}
\title{Cohort Component Method of Population Projection (CCMPP)}
\usage{
ccmpp(
  inputs,
  settings,
  value_col = "value",
  assert_positive_pop = TRUE,
  validate_arguments = TRUE,
  gen_end_interval_col = TRUE
)
}
\arguments{
\item{inputs}{[\code{list()}]\cr
[\code{data.table()}] for each ccmpp input. Requires 'srb', 'asfr', 'baseline';
mortality estimates provided as just 'survival' or two of 'mx', 'ax', and
'qx'; and migration estimates provided as just 'net_migration' or both
'immigration' and 'emigration'. See \strong{Section: Inputs} for more
information on each of the required inputs.}

\item{settings}{[\code{list()}]\cr
Named list of settings for running \code{\link[=ccmpp]{ccmpp()}} with. See
\strong{Section: Settings} for more information on each of the required
settings.}

\item{value_col}{[\code{character(1)}]\cr
Name of the column containing the value of interest in each of the
\code{inputs}. Default is 'value'.}

\item{assert_positive_pop}{[\code{logical(1)}]\cr
Whether or not to check that the projected population estimates are all
greater than or equal to zero. Default is 'TRUE'.}

\item{validate_arguments}{[\code{logical(1)}]\cr
Whether to validate that the input arguments are formatted correctly.
Default is 'TRUE'.}

\item{gen_end_interval_col}{[\code{logical(1)}]\cr
Whether to automatically generate the end of the interval columns (like
'age'). Default is 'TRUE'.}
}
\value{
[\code{data.table()}] of year-sex-age specific population counts.
}
\description{
Implements the deterministic female dominant cohort component
method of population projection.
}
\details{
Given information on the starting population size, the population at
later time points is a function of mortality, fertility, sex-ratio at birth,
and net-migration.

The cohort component method of population projection (CCMPP) projects a
sex-age specific population forward in time using leslie matrices
\code{\link[=leslie_matrix]{leslie_matrix()}} which encode information about fertility, sex-ratio at
birth, and mortality.

\deqn{Population(t + \delta) = Leslie[t, t + \delta] * Population(t)}

See \code{vignette("ccmpp", package = "demCore")} or linked references for more
details on this method.

For mortality related inputs, supplying 'survival' only, or two or more of
'mx', 'qx', and 'ax' is possible. When 'mx', 'qx', and 'ax' parameters are
given then the function calculates the survivorship ratio with
\code{\link[=nSx_from_lx_nLx_Tx]{nSx_from_lx_nLx_Tx()}}.
}
\section{Inputs}{

srb: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item \code{value_col}: [\code{numeric()}] sex-ratio at birth estimates, must be
greater than zero.
}

asfr: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages_asfr' setting.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual age-specific fertility rate
estimates, must be greater than zero.
}

baseline: [\code{data.table()}]\cr
\itemize{
\item year: [\code{integer()}] mid-year for population estimate.
Corresponds to 'years' \code{setting}.
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] baseline year population count estimates,
must be greater than zero.
}

survival: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] survivorship ratio estimates, must be
greater than zero and less than one.
}

mx: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] mortality rate estimates, must be greater
than zero.
}

ax: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] average years lived by those dying in the
interval estimates, must be greater than zero and less than the age
interval length.
}

qx: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] probability of death estimates, must be
greater than zero and less than one.
}

net_migration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual net-migration proportion estimates.
}

immigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual immigration proportion estimates,
must be greater than zero.
}

emigration: [\code{data.table()}]\cr
\itemize{
\item year_start: [\code{integer()}] start of the calendar year interval
(inclusive). Corresponds to 'years' \code{setting}.
\item year_end: [\code{integer()}] end of the calendar year interval (exclusive).
\item sex: [\code{character()}] either 'female' or 'male'. Corresponds to 'sexes'
\code{setting}.
\item age_start: [\code{integer()}] start of the age group (inclusive).
Corresponds to 'ages' \code{setting}.
\item age_end: [\code{integer()}] end of the age group (exclusive).
\item \code{value_col}: [\code{numeric()}] annual emigration proportion estimates, must
be greater than zero.
}
}

\section{Settings}{

\itemize{
\item years: [\code{numeric()}]\cr
The start of each calendar year interval to project in \code{\link[=ccmpp]{ccmpp()}}.
Corresponds to the 'year_start' column in each of the year-specific inputs.
\item sexes: [\code{character()}]\cr
The sexes being projected in \code{\link[=ccmpp]{ccmpp()}}, must be either just
'female', or 'female' and 'male'. Corresponds to the 'sex' column in each
of the sex-specific inputs.
\item ages: [\code{numeric()}]\cr
The ages being projected in \code{\link[=ccmpp]{ccmpp()}}. Corresponds to the
'age_start' column in each of the standard age-specific inputs.
\item ages_mortality: [\code{numeric()}]\cr
The ages for which mortality parameter estimates are available, includes one
extra age group compared to the 'ages' setting. Corresponds to the
'age_start' column in the mortality [\code{data.table()}] input(s).
\item ages_asfr: [\code{numeric()}]\cr
The assumed female reproductive ages, subset of 'ages'. Corresponds to the
'age_start' column in the age-specific fertility rate (asfr)
[\code{data.table()}] input.
}
}

\examples{
population <- ccmpp(
  inputs = thailand_initial_estimates,
  settings = list(
    years = seq(1960, 1995, 5),
    sexes = c("female", "male"),
    ages = seq(0, 80, 5),
    ages_mortality = seq(0, 85, 5),
    ages_asfr = seq(15, 45, 5)
  )
)

}
\references{
Preston, Samuel, Patrick Heuveline, and Michel Guillot. 2001. Demography:
Measuring and Modeling Population. Wiley.
}
\seealso{
\code{\link[=leslie_matrix]{leslie_matrix()}}

vignette("ccmpp")
}
