% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship_ratio.R
\name{determine_age_int}
\alias{determine_age_int}
\title{Helper function to identify unique age interval for input data.table}
\usage{
determine_age_int(dt)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data that includes columns for \code{id_cols}, 'lx', 'nLx' and 'Tx'.}
}
\value{
[\code{numeric(1)}] unique age interval
}
\description{
Identifies age intervals present excluding the terminal age
group and asserts it is uniform across the input data.table.
}
\keyword{internal}
