% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleLT}
\alias{exampleLT}
\title{Example life table}
\format{[\code{data.table()}] columns:
\describe{
\item{age_start}{[\code{integer()}]\cr Abridged ages 0, 1, 5, 10, ..., 95}
\item{age_end}{[\code{integer()}]\cr End of age group inclusive, 1, 5, 10,
... , 95, Inf}
\item{deaths}{[\code{numeric()}]\cr Death counts}
\item{population}{[\code{numeric()}]\cr Population counts}
\item{mx}{[\code{numeric()}]\cr Mortality rate (deaths / population)}
\item{ax}{[\code{numeric()}]\cr Mean person-years lived by those who die in
interval. Obtained using \code{mx_to_ax} function.}
\item{qx}{[\code{numeric()}]\cr Probability of death in interval conditional
on survival to xth birthday. Obtained using \code{mx_ax_to_qx} function.}
}}
\usage{
exampleLT
}
\description{
Example death count and population values. Also
includes mx (deaths / population), ax, and qx calculated using functions
from this package.
}
\examples{
exampleLT
}
\keyword{datasets}
