% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lifetable_parameters.R
\name{gen_lifetable_parameters}
\alias{gen_lifetable_parameters}
\alias{qx_to_lx}
\alias{lx_to_qx}
\alias{lx_to_dx}
\alias{gen_nLx}
\alias{gen_Tx}
\alias{gen_ex}
\title{Generate life table parameters}
\usage{
qx_to_lx(dt, id_cols, assert_na = T)

lx_to_qx(dt, id_cols, terminal_age = 110, assert_na = T)

lx_to_dx(dt, id_cols, terminal_age = 110, assert_na = T)

gen_nLx(dt, id_cols, terminal_age = 110, assert_na = T)

gen_Tx(dt, id_cols, assert_na = T)

gen_ex(dt, assert_na = T)
}
\arguments{
\item{dt}{[\code{data.table()}] variables lx, ax, dx, mx, age, and age_length.}

\item{id_cols}{[\code{character()}] columns that uniquely identify each row
of \code{dt}.}

\item{assert_na}{[\code{logical()}] whether to check for NA values in the
generated variable.}

\item{terminal_age}{[\code{integer(1)}] the terminal age group for the data.
Default: 110.}
}
\value{
dt with column added for new life table parameter. Modifies
data.tables in place.
}
\description{
These functions perform life table calculations, to add selected life table
parameters from others, using established demographic relationships.

The following are required inputs for given parameters, as written in
these functions:
\itemize{
\item dx requires lx
\item lx requires qx
\item qx requires lx
\item ex requires lx and Tx
\item Tx requires nLx
\item nLx requires lx, ax, and dx
}
}
\details{
\strong{Parameter definitions:}
For an age interval x to x+n:
\itemize{
\item mx = mortality rate (deaths / person-years)
\item qx = probability of death, conditional on survival to x-th birthday
\item ax = average years lived of those who died in the age interval
\item dx = number (or proportion of cohort) who died in the age interval
\item lx = number (or proportion of cohort) alive at the beginning of the age
interval
\item ex = life expectancy, expected value for years lived after age x
\item Tx =  total person-years lived beyond age x
\item nLx = total person-years lived in interval
}

\strong{qx_to_lx:} We can use probability of death (qx) to get the proportion of
survivors in the beginning of an age group (lx).
\deqn{At age = 0: lx = 1}
\deqn{At age > 0: lx = lx_previous * (1-qx_previous)}

\strong{lx_to_qx:} We can use survival proportion (lx) and survival proportion at
next age group (l{x+n}) to get probability of death (qx).
\deqn{qx = 1 - (l{x+n}/lx)}
\deqn{For terminal age group: qx = 1}

\strong{lx_to_dx:} Given the proportion surviving to each age group (lx), we can
calculate the proportion dying in the age interval (dx).
\deqn{dx = lx - l{x+n}}
\deqn{For terminal age group: dx = lx}

\strong{gen_nLx:} To get \eqn{nLx} we calculate years lived by survivors plus years
lived by those who died. For terminal ages, assume that average person years
lived equals number at the start of the age group divided by the death rate.
\deqn{nLx = n * l{x+n} + ax * d_x}
\deqn{For terminal age group: infinityLx = lx / mx}

\strong{gen_Tx:} Person-years lived above age x (Tx) is based on person-years
lived by people in each age interval above age x (nLx).
\deqn{For terminal age group: Tx = nLx}
\deqn{Otherwise: Tx = cumulative sum of nLx beyond age x}

\strong{gen_ex:} Life expectancy (ex) at age x is based on person-years lived
above the age group (Tx) and proportion of people alive at the start of
age group (lx).
\deqn{ex = Tx / lx}
}
\examples{
dt <- data.table::data.table(
  sex = rep("both", 4),
  age = c(0, 5, 10, 15),
  age_length = c(5, 5, 5, 120),
  mx = c(0.1, 0.2, 0.3, 0.4),
  ax = c(2.5, 2.5, 2.5, 2.5)
)
dt[, qx := mx_ax_to_qx(mx, ax, age_length)]
qx_to_lx(dt, id_cols = c("sex", "age"))
lx_to_dx(dt, id_cols = c("sex", "age"), terminal_age = 15)
gen_nLx(dt, id_cols = c("sex", "age"), terminal_age = 15)
gen_Tx(dt, id_cols = c("sex", "age"))
gen_ex(dt)
lx_to_qx(dt, id_cols = c("sex", "age"), terminal_age = 15)

}
\seealso{
lifetableUtils::mx_qx_ax_conversions, lifetableUtils::lifetable
}
