% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcab.R
\name{macb_from_nfx}
\alias{macb_from_nfx}
\title{Calculate Mean Age of Childbearing from Fertility Rate (MACB)}
\usage{
macb_from_nfx(dt, id_cols, nfx_col = "nfx", value_col = "macb")
}
\arguments{
\item{dt}{[\code{data.table()}]\cr A data.table with columns 'age_start',
'age_end', and a column of fertility rates named after \code{nfx_col}.}

\item{id_cols}{[\code{character()}]\cr Columns that uniquely identify each row
of \code{dt}. Must include 'age_start' and 'age_end'.}

\item{nfx_col}{[\code{character(1)}]\cr Name of fertility rate column in \code{dt}.
Defualts to "nfx".}

\item{value_col}{[\code{character(1)}]\cr Name of output 'MACB' column. Defaults
to "mcab"}
}
\value{
[\code{data.table()}] A data.table with calculated 'MACB' for each
unique grouping in \code{id_cols}, stored in \code{value_col}.
}
\description{
The mean age at childbearing is the mean age of mothers at the birth of their
children if women were subject throughout their lives to the age-specific
fertility rates observed in a given year.

See \href{https://www.un.org/en/development/desa/population/publications/dataset/fertility/age-childbearing.asp}{UNPOP page} for more deatils.
}
\examples{
asfr_data <- data.table(
  asfr = c(.051, .196, .208, .147, .075, .024, .004),
  age_start = seq(15, 45, 5),
  age_end = seq(20, 50, 5)
 )

id_cols <- c("age_start", "age_end")
macb_dt <- macb_from_nfx(asfr_data, id_cols = id_cols, nfx_col = "asfr")


}
