% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_qx_ax_conversions.R
\name{mx_qx_ax_conversions}
\alias{mx_qx_ax_conversions}
\alias{mx_to_qx}
\alias{mx_ax_to_qx}
\alias{qx_to_mx}
\alias{mx_qx_to_ax}
\alias{qx_ax_to_mx}
\alias{mx_to_ax}
\title{Convert between mx, qx, and ax values}
\usage{
mx_to_qx(mx, length)

mx_ax_to_qx(mx, ax, length)

qx_to_mx(qx, length)

mx_qx_to_ax(mx, qx, length)

qx_ax_to_mx(qx, ax, length)

mx_to_ax(mx, length)
}
\arguments{
\item{mx}{[\code{numeric()}] mx (mortality death rate)}

\item{length}{[\code{numeric()}] the length of the age interval (e.g. 5 for
age 5-9)}

\item{ax}{[\code{numeric()}] ax (average number of years lived in the age
interval of those who died)}

\item{qx}{[\code{numeric()}] qx (probability of death)}
}
\value{
Numeric value representing the converted mx, qx, or ax value.
}
\description{
Convert between mx and qx values given a time for an age
interval. If ax (the average number of years lived in the x to x + n age
interval by those dying in the interval) exists, use the \emph{\emph{ax_to}}
functions. The mx_to_qx and qx_to_mx functions assume smooth mortality
across the interval.
}
\examples{
mx_to_qx(0.25, 45)
mx_ax_to_qx(.2, 3, 5)
qx_to_mx(.5, 5)
mx_qx_to_ax(.5, .25, 5)
qx_ax_to_mx(.1, 3, 5)
mx_to_ax(0.2, 5)
}
