% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_popReconstruct.R
\docType{data}
\name{thailand_data}
\alias{thailand_data}
\alias{thailand_initial_estimates}
\title{Thailand example population data and ccmpp initial estimates}
\format{\code{thailand_data}: list containing sex-specific matrices of population
count data.
\itemize{
\item population: named list with female and male matrices of age-specific
census counts in years after the baseline year (1970, 1980, 1990, 2000).
Rows correspond to age groups while columns correspond to years.
}

\code{thailand_initial_estimates}: list containing matrices of initial
estimates for each \code{\link[=ccmpp]{ccmpp()}} input.
\itemize{
\item srb: single row matrix of initial estimates of sex ratio at birth.
\item asfr: matrix of initial estimates of average annual single-year
age-specific fertility rates.
\item baseline: List with female and male single column matrices of
age-specific population counts in the baseline year (1960).
\item survival: List with female and male matrices of initial estimates of
age-specific survivorship ratios
\item net_migration: List with female and male matrices of initial estimates of
age-specific average annual net migration proportions.
}}
\usage{
thailand_data

thailand_initial_estimates
}
\description{
Example data and initial estimates for reconstructing the female
and male population in Thailand from 1960 to 2000.
}
\references{
Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland.
2015. “Bayesian Reconstruction of Two-Sex Populations by Age: Estimating Sex
Ratios at Birth and Sex Ratios of Mortality.” Journal of the Royal
Statistical Society. Series A: Statistics in Society 178 (4): 977–1007.
\url{https://doi.org/10.1111/rssa.12104}.

\href{https://github.com/markalava/Bayesian-Reconstruction}{markalava/Bayesian-Reconstruction github repo}
}
\seealso{
\code{\link[=ccmpp]{ccmpp()}}, \code{\link[populationMethods:popReconstruct]{populationMethods::popReconstruct()}}
}
\keyword{datasets}
