% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmpp.R
\name{validate_ccmpp_inputs}
\alias{validate_ccmpp_inputs}
\title{Validate the ccmpp input arguments}
\usage{
validate_ccmpp_inputs(inputs, input_years, input_sexes, input_ages, value_col)
}
\arguments{
\item{inputs}{list of [\code{data.table()}]s of initial estimates for each ccmpp
input.
\itemize{
\item srb: [\code{data.table()}] of year-specific sex ratio at birth estimates.
\item asfr: [\code{data.table()}] of year-age-specific average annual single-year
age-specific fertility rate estimates. Must include both maternal and
non-maternal age groups that are zero.
\item baseline: [\code{data.table()}] of year-sex-age specific baseline year
population counts.
\item survival: [\code{data.table()}] of year-sex-age-specific survivorship ratio
estimates.
\item net_migration: [\code{data.table()}] of year-sex-age-specific average annual
net migration proportion estimates
}}

\item{input_years}{[\code{numeric()}]\cr
the unique 'year_start' column values in each of the \code{inputs}. The minimum
value must correspond to the only 'year_start' value in the baseline input.}

\item{input_sexes}{[\code{character()}]\cr
the unique 'sex' column values in each of the sex-specific \code{inputs}. Must
be either just 'female', or 'female' and 'male'.}

\item{input_ages}{[\code{numeric()}]\cr
the unique 'age_start' column values in each of the age-specific \code{inputs}.}

\item{value_col}{[\code{character(1)}]\cr
Name of the column containing the value of interest in each of the
\code{inputs}. Default is 'value'.}
}
\value{
Invisibly returns \code{inputs} but throws error if not formatted
correctly.
}
\description{
Validate the ccmpp input arguments
}
